% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{get_species_categories}
\alias{get_species_categories}
\title{Get valid NWRSpecies taxonomic categories}
\usage{
get_species_categories(api_version = .nwrs_api_version)
}
\arguments{
\item{api_version}{Character scalar. Required API version.}
}
\value{
A tibble of category names.
}
\description{
Queries the \code{SpeciesList/categories} endpoint for the category names
accepted by \code{\link[=get_species_list]{get_species_list()}}.
}
\examples{
\dontrun{
get_species_categories()
}
}
